Move trust wallet into its own composable
